Enum mistralrs_server::model_selected::ModelSelected

pub enum ModelSelected {

    Mistral {
        model_id: String,
        tokenizer_json: Option<String>,
        repeat_last_n: usize,
    },
    MistralGGUF {
        tok_model_id: String,
        tokenizer_json: Option<String>,
        quantized_model_id: Option<String>,
        quantized_filename: Option<String>,
        repeat_last_n: usize,
    },
    XLoraMistral {
        model_id: String,
        tokenizer_json: Option<String>,
        xlora_model_id: String,
        repeat_last_n: usize,
        order: String,
        tgt_non_granular_index: Option<usize>,
    },
    Gemma {
        model_id: String,
        tokenizer_json: Option<String>,
        repeat_last_n: usize,
    },
    XLoraGemma {
        model_id: String,
        tokenizer_json: Option<String>,
        xlora_model_id: String,
        repeat_last_n: usize,
        order: String,
        tgt_non_granular_index: Option<usize>,
    },
    Llama {
        model_id: String,
        tokenizer_json: Option<String>,
        repeat_last_n: usize,
    },
    LlamaGGUF {
        tok_model_id: String,
        tokenizer_json: Option<String>,
        quantized_model_id: Option<String>,
        quantized_filename: Option<String>,
        repeat_last_n: usize,
    },
    LlamaGGML {
        tok_model_id: String,
        tokenizer_json: Option<String>,
        quantized_model_id: Option<String>,
        quantized_filename: Option<String>,
        repeat_last_n: usize,
        gqa: usize,
    },
    XLoraLlama {
        model_id: String,
        tokenizer_json: Option<String>,
        xlora_model_id: String,
        repeat_last_n: usize,
        order: String,
        tgt_non_granular_index: Option<usize>,
    },
    Mixtral {
        model_id: String,
        tokenizer_json: Option<String>,
        repeat_last_n: usize,
    },
    MixtralGGUF {
        tok_model_id: String,
        tokenizer_json: Option<String>,
        quantized_model_id: Option<String>,
        quantized_filename: Option<String>,
        repeat_last_n: usize,
    },
    XLoraMixtral {
        model_id: String,
        tokenizer_json: Option<String>,
        xlora_model_id: String,
        repeat_last_n: usize,
        order: String,
        tgt_non_granular_index: Option<usize>,
    },
    XLoraMistralGGUF {
        tok_model_id: String,
        tokenizer_json: Option<String>,
        quantized_model_id: Option<String>,
        quantized_filename: Option<String>,
        repeat_last_n: usize,
        xlora_model_id: String,
        order: String,
        tgt_non_granular_index: Option<usize>,
    },
    XLoraLlamaGGUF {
        tok_model_id: String,
        tokenizer_json: Option<String>,
        quantized_model_id: Option<String>,
        quantized_filename: Option<String>,
        repeat_last_n: usize,
        xlora_model_id: String,
        order: String,
        tgt_non_granular_index: Option<usize>,
    },
    XLoraLlamaGGML {
        tok_model_id: String,
        tokenizer_json: Option<String>,
        quantized_model_id: Option<String>,
        quantized_filename: Option<String>,
        repeat_last_n: usize,
        xlora_model_id: String,
        order: String,
        gqa: usize,
        tgt_non_granular_index: Option<usize>,
    },
    XLoraMixtralGGUF {
        tok_model_id: String,
        tokenizer_json: Option<String>,
        quantized_model_id: Option<String>,
        quantized_filename: Option<String>,
        repeat_last_n: usize,
        xlora_model_id: String,
        order: String,
        tgt_non_granular_index: Option<usize>,
    },
    Phi2 {
        model_id: String,
        tokenizer_json: Option<String>,
        repeat_last_n: usize,
    },

}

Variants

Mistral

Fields

model_id: String

Model ID to load from

tokenizer_json: Option<String>

Path to local tokenizer.json file. If this is specified it is used over any remote file.

repeat_last_n: usize

Control the application of repeat penalty for the last n tokens

Select the mistral model.

MistralGGUF

Fields

tok_model_id: String

Model ID to load the tokenizer from

tokenizer_json: Option<String>

Path to local tokenizer.json file. If this is specified it is used over any remote file.

quantized_model_id: Option<String>

Quantized model ID to find the quantized_filename, only applicable if quantized is set. If it is set to an empty string then the quantized filename will be used as a path to the GGUF file.

quantized_filename: Option<String>

Quantized filename, only applicable if quantized is set.

repeat_last_n: usize

Control the application of repeat penalty for the last n tokens

Select the quantized mistral model with gguf.

XLoraMistral

Fields

model_id: String

Model ID to load from

tokenizer_json: Option<String>

Path to local tokenizer.json file. If this is specified it is used over any remote file.

xlora_model_id: String

Model ID to load Xlora from

repeat_last_n: usize

Control the application of repeat penalty for the last n tokens

order: String

Ordering JSON file

tgt_non_granular_index: Option<usize>

Index of completion tokens to generate scalings up until. If this is 1, then there will be one completion token generated before it is cached.

Select the mistral model, with X-LoRA.

Gemma

Fields

model_id: String

Model ID to load from

tokenizer_json: Option<String>

Path to local tokenizer.json file. If this is specified it is used over any remote file.

repeat_last_n: usize

Control the application of repeat penalty for the last n tokens

Select the gemma model.

XLoraGemma

Fields

model_id: String

Model ID to load from

tokenizer_json: Option<String>

Path to local tokenizer.json file. If this is specified it is used over any remote file.

xlora_model_id: String

Model ID to load Xlora from

repeat_last_n: usize

Control the application of repeat penalty for the last n tokens

order: String

Ordering JSON file

tgt_non_granular_index: Option<usize>

Index of completion tokens to generate scalings up until. If this is 1, then there will be one completion token generated before it is cached.

Select the gemma model, with X-LoRA.

Llama

Fields

model_id: String

Model ID to load from

tokenizer_json: Option<String>

Path to local tokenizer.json file. If this is specified it is used over any remote file.

repeat_last_n: usize

Control the application of repeat penalty for the last n tokens

Select the llama model.

LlamaGGUF

Fields

tok_model_id: String

Model ID to load the tokenizer from

tokenizer_json: Option<String>

Path to local tokenizer.json file. If this is specified it is used over any remote file.

quantized_model_id: Option<String>

Quantized model ID to find the quantized_filename, only applicable if quantized is set. If it is set to an empty string then the quantized filename will be used as a path to the GGUF file.

quantized_filename: Option<String>

Quantized filename, only applicable if quantized is set.

repeat_last_n: usize

Control the application of repeat penalty for the last n tokens

Select the quantized llama model with gguf.

LlamaGGML

Fields

tok_model_id: String

Model ID to load the tokenizer from

tokenizer_json: Option<String>

Path to local tokenizer.json file. If this is specified it is used over any remote file.

quantized_model_id: Option<String>

Quantized model ID to find the quantized_filename, only applicable if quantized is set. If it is set to an empty string then the quantized filename will be used as a path to the GGML file.

quantized_filename: Option<String>

Quantized filename, only applicable if quantized is set.

repeat_last_n: usize

Control the application of repeat penalty for the last n tokens

gqa: usize

GQA

Select the quantized llama model with gguf.

XLoraLlama

Fields

model_id: String

Model ID to load from

tokenizer_json: Option<String>

Path to local tokenizer.json file. If this is specified it is used over any remote file.

xlora_model_id: String

Model ID to load Xlora from

repeat_last_n: usize

Control the application of repeat penalty for the last n tokens

order: String

Ordering JSON file

tgt_non_granular_index: Option<usize>

Index of completion tokens to generate scalings up until. If this is 1, then there will be one completion token generated before it is cached.

Select the llama model, with X-LoRA.

Mixtral

Fields

model_id: String

Model ID to load from

tokenizer_json: Option<String>

Path to local tokenizer.json file. If this is specified it is used over any remote file.

repeat_last_n: usize

Control the application of repeat penalty for the last n tokens

Select the mixtral model.

MixtralGGUF

Fields

tok_model_id: String

Model ID to load the tokenizer from

tokenizer_json: Option<String>

Path to local tokenizer.json file. If this is specified it is used over any remote file.

quantized_model_id: Option<String>

Quantized model ID to find the quantized_filename, only applicable if quantized is set. If it is set to an empty string then the quantized filename will be used as a path to the GGUF file.

quantized_filename: Option<String>

Quantized filename, only applicable if quantized is set.

repeat_last_n: usize

Control the application of repeat penalty for the last n tokens

Select the quantized mixtral model with gguf.

XLoraMixtral

Fields

model_id: String

Model ID to load from

tokenizer_json: Option<String>

Path to local tokenizer.json file. If this is specified it is used over any remote file.

xlora_model_id: String

Model ID to load Xlora from

repeat_last_n: usize

Control the application of repeat penalty for the last n tokens

order: String

Ordering JSON file

tgt_non_granular_index: Option<usize>

Index of completion tokens to generate scalings up until. If this is 1, then there will be one completion token generated before it is cached.

Select the mixtral model, with X-LoRA.

XLoraMistralGGUF

Fields

tok_model_id: String

Model ID to load the tokenizer from

tokenizer_json: Option<String>

Path to local tokenizer.json file. If this is specified it is used over any remote file.

quantized_model_id: Option<String>

Quantized model ID to find the quantized_filename, only applicable if quantized is set. If it is set to an empty string then the quantized filename will be used as a path to the GGUF file.

quantized_filename: Option<String>

Quantized filename, only applicable if quantized is set.

repeat_last_n: usize

Control the application of repeat penalty for the last n tokens

xlora_model_id: String

Model ID to load Xlora from

order: String

Ordering JSON file

tgt_non_granular_index: Option<usize>

Index of completion tokens to generate scalings up until. If this is 1, then there will be one completion token generated before it is cached.

Select the quantized mistral model with gguf and X-LoRA.

XLoraLlamaGGUF

Fields

tok_model_id: String

Model ID to load the tokenizer from

tokenizer_json: Option<String>

Path to local tokenizer.json file. If this is specified it is used over any remote file.

quantized_model_id: Option<String>

Quantized model ID to find the quantized_filename, only applicable if quantized is set. If it is set to an empty string then the quantized filename will be used as a path to the GGUF file.

quantized_filename: Option<String>

Quantized filename, only applicable if quantized is set.

repeat_last_n: usize

Control the application of repeat penalty for the last n tokens

xlora_model_id: String

Model ID to load Xlora from

order: String

Ordering JSON file

tgt_non_granular_index: Option<usize>

Index of completion tokens to generate scalings up until. If this is 1, then there will be one completion token generated before it is cached.

Select the quantized mistral model with gguf and X-LoRA.

XLoraLlamaGGML

Fields

tok_model_id: String

Model ID to load the tokenizer from

tokenizer_json: Option<String>

Path to local tokenizer.json file. If this is specified it is used over any remote file.

quantized_model_id: Option<String>

Quantized model ID to find the quantized_filename, only applicable if quantized is set. If it is set to an empty string then the quantized filename will be used as a path to the GGML file.

quantized_filename: Option<String>

Quantized filename, only applicable if quantized is set.

repeat_last_n: usize

Control the application of repeat penalty for the last n tokens

xlora_model_id: String

Model ID to load Xlora from

order: String

Ordering JSON file

gqa: usize

GQA

tgt_non_granular_index: Option<usize>

Index of completion tokens to generate scalings up until. If this is 1, then there will be one completion token generated before it is cached.

Select the quantized mistral model with gguf and X-LoRA.

XLoraMixtralGGUF

Fields

tok_model_id: String

Model ID to load the tokenizer from

tokenizer_json: Option<String>

Path to local tokenizer.json file. If this is specified it is used over any remote file.

quantized_model_id: Option<String>

Quantized model ID to find the quantized_filename, only applicable if quantized is set. If it is set to an empty string then the quantized filename will be used as a path to the GGUF file.

quantized_filename: Option<String>

Quantized filename, only applicable if quantized is set.

repeat_last_n: usize

Control the application of repeat penalty for the last n tokens

xlora_model_id: String

Model ID to load Xlora from

order: String

Ordering JSON file

tgt_non_granular_index: Option<usize>

Index of completion tokens to generate scalings up until. If this is 1, then there will be one completion token generated before it is cached.

Select the quantized mistral model with gguf and X-LoRA.

Phi2

Fields

model_id: String

Model ID to load from

tokenizer_json: Option<String>

Path to local tokenizer.json file. If this is specified it is used over any remote file.

repeat_last_n: usize

Control the application of repeat penalty for the last n tokens

Select the phi2 model.